requirements.txt
fastapi
uvicorn[standard]
requests
pyyaml
python-dotenv

